refactor(suite): rename TradingFormInputAssetPicker to TradingFormInputBuyAsset and extract common to TradingFormInputAssetPicker
