+++ packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputAssetPicker/index.ts
@@ -1,0 +1,2 @@
+export * from './AssetPickerInput/AssetPickerInput';
+export * from './AssetSearchWithNetworkFilter/AssetSearchWithNetworkFilter';
