--- packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputAssetPicker/types/index.ts
@@ -1,3 +1,0 @@
-import { TradingBuyFormProps, TradingExchangeFormProps } from '@suite-common/trading';
-
-export type TradingFormValues = TradingExchangeFormProps | TradingBuyFormProps;
